Component for BaseSaveButtonWithActionMenu
